requirements.txt
requests
pymongo
apache-airflow